+++ src/modelFactory/alignmentFactory.ts
@@ -1,23 +1,19 @@
 import { Alignment } from "../models/alignment";
 
 export class AlignmentFactory {
+
     public createAlignments(cells: string[]): Alignment[] {
-        let result: Alignment[] = [];
+        return cells.map(this.alignmentOf);
+    }
 
-        const len: number = cells.length;
-        for (let i = 0; i < len; i++) {
-            const cell = cells[i];
-            const left = cell[0] == ":";;
-            const right = cell[cell.length - 1] == ":";
+    private alignmentOf(cell: string): Alignment {
+        const left = cell[0] == ":";
+        const right = cell[cell.length - 1] == ":";
 
-            if (left && right)
-                result.push(Alignment.Center);
-            else if (right)
-                result.push(Alignment.Right);
-            else
-                result.push(Alignment.Left);
-        }
+        if (left && right) return Alignment.Center;
+        if (right) return Alignment.Right;
+        if (left) return Alignment.Left;
 
-        return result;
+        return Alignment.NotSet;
     }
 }