+++ src/padCalculation/center/lastColumnPadCalculator.ts
@@ -1,0 +1,9 @@
+import { Table } from "../../models/table";
+import { CenterPadCalculator } from "./centerPadCalculator";
+
+export class LastColumnPadCalculator extends CenterPadCalculator {
+    public getRightPadding(paddingChar: string, table: Table, row: number, column: number): string {
+        if (!table.hasRightBorder) return "";
+        return super.getRightPadding(paddingChar, table, row, column);
+    }
+}