+++ src/padCalculation/left/middleColumnPadCalculator.ts
@@ -1,9 +1,9 @@
-import { BasePadCalculator } from "./basePadCalculator";
-import { Table } from "../models/table";
-import { Cell } from "../models/cell";
+import { BasePadCalculator } from "../basePadCalculator";
+import { Table } from "../../models/table";
+import { Cell } from "../../models/cell";
 
 export class MiddleColumnPadCalculator extends BasePadCalculator {
-    public getLeftPadding(paddingChar: string, table: Table, cell: Cell): string {
+    public getLeftPadding(paddingChar: string, table: Table, row: number, column: number): string {
         return paddingChar;
     }
     public getRightPadding(paddingChar: string, table: Table, row: number, column: number): string {