+++ src/padCalculation/right/middleColumnPadCalculator.ts
@@ -1,0 +1,7 @@
+import { Table } from "../../models/table";
+import { Cell } from "../../models/cell";
+import { RightPadCalculator } from "./rightPadCalculator";
+
+export class MiddleColumnPadCalculator extends RightPadCalculator {
+
+}