+++ test/systemTests/tableRangePrettyfierFactory.ts
@@ -2,7 +2,6 @@ import * as assert from 'assert';
 import * as vscode from "vscode";
 import * as fs from 'fs';
 import * as path from 'path';
-import { Mock } from "typemoq";
 import { TableRangePrettyfier } from "../../src/extension/tableRangePrettyfier";
 import { TableFactory } from "../../src/modelFactory/tableFactory";
 import { AlignmentFactory } from "../../src/modelFactory/alignmentFactory";
@@ -17,8 +16,8 @@ import { MarkdownTextDocumentStub } from "../stubs/markdownTextDocumentStub";
 import { TrimmerTransformer } from '../../src/modelFactory/transformers/trimmerTransformer';
 import { BorderTransformer } from '../../src/modelFactory/transformers/borderTransformer';
 import { SelectionInterpreter } from '../../src/modelFactory/selectionInterpreter';
-import { SeparatorPadCalculator } from '../../src/padCalculation/separatorPadCalculator';
 import { PadCalculatorSelector } from '../../src/padCalculation/padCalculatorSelector';
+import { AlignmentMarkerStrategy } from '../../src/viewModelFactories/alignmentMarking';
 
 export class PrettyfierFromFile {
     private readonly _logger: ILogger;
@@ -62,7 +61,7 @@ export class PrettyfierFromFile {
             new TableValidator(new SelectionInterpreter()),
             new TableViewModelFactory(new RowViewModelFactory(
                 new ContentPadCalculator(new PadCalculatorSelector(), " "), 
-                new SeparatorPadCalculator(new PadCalculatorSelector(), "-")
+                new AlignmentMarkerStrategy(":")
             )),
             new TableStringWriter(),
             [ this._logger ]