+++ test/unitTests/extension/extension.test.ts
@@ -11,7 +11,7 @@ suite("Extension Tests", () => {
     test("Extension gets activated", () => {
          vscode.extensions.getExtension(_extensionName)
             .activate()
-            .then((publicApi) => {
+            .then(() => {
                     assert.ok(true);
                 }, rejectReason => {
                     assert.fail(0, 1, `Extension not activated. Reason: ${rejectReason}`, null);