+++ test/unitTests/padCalculation/contentPadCalculator.test.ts
@@ -1,5 +1,4 @@
 import * as assert from 'assert';
-import { assertExt } from "../../assertExtensions";
 import { IMock, Mock, It, Times } from 'typemoq';
 import { ContentPadCalculator } from "../../../src/padCalculation/contentPadCalculator";
 import { Table } from '../../../src/models/table';
@@ -26,7 +25,7 @@ suite("ContentPadCalculator tests", () => {
             [ "aaaaa", "bbbbb", "ccccc" ]
         ]);
         _selector.setup(_ => _.select(table, It.isAny())).returns(() => _mockCalculator.object).verifiable(Times.once());
-        _mockCalculator.setup(_ => _.getLeftPadding(It.isAny(), table, It.isAny())).returns(() => "test").verifiable(Times.once());
+        _mockCalculator.setup(_ => _.getLeftPadding(It.isAny(), table, It.isAny(), It.isAny())).returns(() => "test").verifiable(Times.once());
 
         const pad = sut.getLeftPadding(table, 1, 2);
 
@@ -52,7 +51,7 @@ suite("ContentPadCalculator tests", () => {
     });
 
     function tableFor(rows: string[][]) {
-        const alignments: Alignment[] = rows[0].map(r => Alignment.Left);
+        const alignments: Alignment[] = rows[0].map(() => Alignment.Left);
         let table = new Table(rows.map(row => row.map(c  => new Cell(c))), alignments);
         return table;
     }